Refactor the permissions to be a part of the user profile on admin side.
User profile component displays various information about SB users, useful
to admins. Only the information that the admin has access to will be
displayed.
