--- client/admin/admin-utils.js
@@ -1,8 +1,0 @@
-export function isAdmin(authState) {
-  return authState.permissions.valueSeq().some(perm => perm)
-}
-
-export function checkPermissions(authState, ...permissionsToCheck) {
-  const perms = authState.permissions
-  return permissionsToCheck.every(p => perms[p])
-}
