+++ client/admin/admin.css
@@ -1,6 +1,6 @@
 @value colorTextSecondary from '../styles/colors.css';
 
-.permissions {
+.users {
   padding: 0 20px;
 }
