+++ client/admin/panel.jsx
@@ -8,15 +8,15 @@ export default class Panel extends React.Component {
   render() {
     const perms = this.props.auth.permissions
 
-    const permissionsLink = perms.editPermissions ?
-        <li><Link to='/admin/permissions'>Change user permissions</Link></li> : null
+    const usersLink = (perms.editPermissions || perms.banUsers) ?
+        <li><Link to='/admin/users'>View user's profile</Link></li> : null
     const invitesLink = perms.acceptInvites ?
         <li><Link to='/admin/invites'>View beta invites</Link></li> : null
 
     return (
       <ContentLayout title={'Admin panel'}>
         <ul>
-          {permissionsLink}
+          {usersLink}
           {invitesLink}
         </ul>
       </ContentLayout>
